IMPORTANT: changed endpoint functionality for getting tokens
The previous implementation for getting/creating API tokens has been
deprecated - it was causing issues with autodiscovery of views/routes
when generating API schema. In the future, a simple POST request to
the endpoint `/api/token/` with a correct username and password will
return an existing token or create a new one.

You can supply the username/password in two ways:
    - in the request body in JSON
    - in the request body as form-data

An example of JSON would look like this:
```json
{
    "username": "avery",
    "password": "bobbit82"
}
```